Remove artifact (side-effect) of "unfair" voting
With the "12,000", the votes of blocks 97..2016 of each 2016 interval have 20% more weight than the votes of blocks 1..96, because they would influence the next 6 block size limit adjustments instead of only 5. [ 12,000 = 5*2016 + (2016-96) ]
Hence change the constant 12,000 to 6 * 2,016 = 12,096, and adapt the 20% / 80% percentiles accordingly.

The other change is a modification of the phrasing, to define the percentiles more clearly.

Note: I propose to pull https://github.com/jgarzik/bip100/pull/14 instead of this one. Only pull this one if https://github.com/jgarzik/bip100/pull/14 is not accepted.