Parameterization, clarification
Starting hardLimit and activation height are changed to
implementation parameters.

Current hardLimit is clarified to mean the hardLimit of the last
block in the previous interval.  This can matter if some other
mechanism is capable of updating the limit.  This actually
happened with the Bitcoin Cash fork, which occurred in the middle
of a difficulty interval.